arm64 compatibility
This branch is experimental yet until we get a device to test all the classes that have been changed.

+++ Classes/Bindings/Private/CKUIControlBlockBinder.m
@@ -55,8 +55,8 @@
 }
 
 - (NSString*)description{
-	return [NSString stringWithFormat:@"<CKUIControlBlockBinder : %p>{\ncontrolRef = %@\ncontrolEvents = %d}",
-			self,self.control ? self.control : @"(null)",controlEvents];
+	return [NSString stringWithFormat:@"<CKUIControlBlockBinder : %p>{\ncontrolRef = %@\ncontrolEvents = %lu}",
+			self,self.control ? self.control : @"(null)",(unsigned long)controlEvents];
 }
 
 - (void)reset{
