+++ Classes/Foundation/Public/Helpers/NSString+Formating.m
@@ -19,7 +19,7 @@
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@"-" withString:@""];
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@"+" withString:@""];
     
-    int length = [mobileNumber length];
+    NSUInteger length = [mobileNumber length];
     if(length > 10)
     {
         mobileNumber = [mobileNumber substringFromIndex: length-10];
@@ -29,7 +29,7 @@
 }
 
 
-+(int)getLength:(NSString*)mobileNumber
++(NSUInteger)getLength:(NSString*)mobileNumber
 {
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@"(" withString:@""];
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@")" withString:@""];
@@ -37,7 +37,7 @@
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@"-" withString:@""];
     mobileNumber = [mobileNumber stringByReplacingOccurrencesOfString:@"+" withString:@""];
     
-    int length = [mobileNumber length];
+    NSUInteger length = [mobileNumber length];
     
     return length;
 }
@@ -54,7 +54,7 @@
     [phoneNumberSet addCharactersInString:@"() _+"];
     NSString* text = [textField.text stringByTrimmingCharactersInSet:[phoneNumberSet invertedSet]];
     
-    int length = [self getLength:text];
+    NSUInteger length = [self getLength:text];
     
     if(length == 10)
     {
@@ -153,7 +153,7 @@
             }
         }
         
-        int length = [text length];
+        NSUInteger length = [text length];
         if(length > 10){
             return NO;
         }
